[REFACTOR] 패키지명 cheerTalk으로 변경 #92
