+++ src/main/java/com/sports/server/command/cheertalk/application/CheerTalkService.java
@@ -1,11 +1,11 @@
-package com.sports.server.command.comment.application;
+package com.sports.server.command.cheertalk.application;
 
-import static com.sports.server.command.comment.exception.CheerTalkErrorMessages.CHEER_TALK_CONTAINS_BAD_WORD;
+import static com.sports.server.command.cheertalk.exception.CheerTalkErrorMessages.CHEER_TALK_CONTAINS_BAD_WORD;
 
-import com.sports.server.command.comment.domain.CheerTalk;
-import com.sports.server.command.comment.domain.CheerTalkRepository;
-import com.sports.server.command.comment.domain.LanguageFilter;
-import com.sports.server.command.comment.dto.CheerTalkRequest;
+import com.sports.server.command.cheertalk.domain.CheerTalk;
+import com.sports.server.command.cheertalk.domain.CheerTalkRepository;
+import com.sports.server.command.cheertalk.domain.LanguageFilter;
+import com.sports.server.command.cheertalk.dto.CheerTalkRequest;
 import com.sports.server.common.exception.CustomException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
