+++ src/main/java/com/sports/server/command/cheertalk/exception/CheerTalkErrorMessages.java
@@ -1,4 +1,4 @@
-package com.sports.server.command.comment.exception;
+package com.sports.server.command.cheertalk.exception;
 
 public class CheerTalkErrorMessages {
     public static final String CHEER_TALK_CONTAINS_BAD_WORD = "댓글에 욕설이 포함되어 있어 저장할 수 없습니다.";
