+++ src/main/java/com/sports/server/query/repository/CheerTalkDynamicRepository.java
@@ -1,6 +1,6 @@
 package com.sports.server.query.repository;
 
-import com.sports.server.command.comment.domain.CheerTalk;
+import com.sports.server.command.cheertalk.domain.CheerTalk;
 
 import java.util.List;
