+++ src/test/java/com/sports/server/command/report/domain/ReportTest.java
@@ -1,6 +1,6 @@
 package com.sports.server.command.report.domain;
 
-import com.sports.server.command.comment.domain.CheerTalk;
+import com.sports.server.command.cheertalk.domain.CheerTalk;
 import com.sports.server.common.exception.CustomException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
