+++ src/test/java/com/sports/server/query/acceptance/CheerTalkEventHandlerTest.java
@@ -3,8 +3,8 @@ package com.sports.server.query.acceptance;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import com.fasterxml.jackson.module.paramnames.ParameterNamesModule;
-import com.sports.server.command.comment.application.CheerTalkService;
-import com.sports.server.command.comment.dto.CheerTalkRequest;
+import com.sports.server.command.cheertalk.application.CheerTalkService;
+import com.sports.server.command.cheertalk.dto.CheerTalkRequest;
 import com.sports.server.query.dto.response.CommentResponse;
 import com.sports.server.support.AcceptanceTest;
 import org.junit.jupiter.api.BeforeEach;
